+++ Src/Private/Get-DiagBackupToProtectedGroup.ps1
@@ -5,7 +5,7 @@ function Get-DiagBackupToProtectedGroup {
     .DESCRIPTION
         Build a diagram of the configuration of Veeam VBR in PDF/PNG/SVG formats using Psgraph.
     .NOTES
-        Version:        0.5.9
+        Version:        0.6.0
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -84,7 +84,7 @@ function Get-DiagBackupToProtectedGroup {
                 $_
             }
 
-            if ($ProtectedGroups) {
+            if ($ProtectedGroups.Container) {
                 if ($Dir -eq 'LR') {
                     $DiagramLabel = 'Protected Groups'
                     $DiagramDummyLabel = ' '
@@ -105,10 +105,10 @@ function Get-DiagBackupToProtectedGroup {
                             Rank DummyPGLeft, DummyPGLeftt, ProtectedGroup, DummyPGRight
                         }
                         if ($ADContainer) {
-                            SubGraph ADContainer -Attributes @{Label = (Get-HTMLLabel -Label 'Active Directory Computers' -IconType "VBR_AGENT_AD_Logo" -SubgraphLabel); fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
+                            SubGraph ADContainer -Attributes @{Label = (Get-DiaHTMLLabel -Label 'Active Directory Computers' -IconType "VBR_AGENT_AD_Logo" -ImagesObj $Images -IconDebug $IconDebug -SubgraphLabel); fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
                                 # Node used for subgraph centering
                                 Node DummyADContainer @{Label = 'DummyADC'; style = $SubGraphDebug.style; color = $SubGraphDebug.color; shape = 'plain' }
-                                if ($ADContainer.count -le 2) {
+                                if (($ADContainer | Measure-Object).count -le 2) {
                                     foreach ($PGOBJ in ($ADContainer | Sort-Object -Property Name)) {
                                         $PGHASHTABLE = @{}
                                         $PGOBJ.psobject.properties | ForEach-Object { $PGHASHTABLE[$_.Name] = $_.Value }
@@ -130,12 +130,12 @@ function Get-DiagBackupToProtectedGroup {
                                             $Ous
                                         )
 
-                                        Convert-TableToHTML -Label $PGOBJ.Name -Name $PGOBJ.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14
+                                        Convert-DiaTableToHTML -Label $PGOBJ.Name -Name $PGOBJ.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14 -IconDebug $IconDebug
 
                                         Edge -From DummyADContainer -To $PGOBJ.Name @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                                     }
                                 } else {
-                                    $Group = Split-Array -inArray ($ADContainer | Sort-Object -Property Name) -size 2
+                                    $Group = Split-array -inArray ($ADContainer | Sort-Object -Property Name) -size 2
                                     $Number = 0
                                     while ($Number -ne $Group.Length) {
                                         $Random = Get-Random
@@ -160,7 +160,7 @@ function Get-DiagBackupToProtectedGroup {
                                                     $Ous
                                                 )
 
-                                                Convert-TableToHTML -Label $_.Name -Name $_.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14
+                                                Convert-DiaTableToHTML -Label $_.Name -Name $_.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14 -IconDebug $IconDebug
                                             }
                                         }
                                         $Number++
@@ -179,10 +179,10 @@ function Get-DiagBackupToProtectedGroup {
                             Edge -From ProtectedGroup -To DummyADContainer @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                         }
                         if ($ManualContainer) {
-                            SubGraph MCContainer -Attributes @{Label = (Get-HTMLLabel -Label 'Manual Computers' -IconType "VBR_AGENT_MC" -SubgraphLabel); fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
+                            SubGraph MCContainer -Attributes @{Label = (Get-DiaHTMLLabel -Label 'Manual Computers' -IconType "VBR_AGENT_MC" -ImagesObj $Images -IconDebug $IconDebug -SubgraphLabel); fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
                                 # Node used for subgraph centering
                                 Node DummyMCContainer @{Label = 'DummyMC'; style = $SubGraphDebug.style; color = $SubGraphDebug.color; shape = 'plain' }
-                                if ($ManualContainer.count -le 2) {
+                                if (($ManualContainer | Measure-Object).count -le 2) {'Backup-to-All'
                                     foreach ($PGOBJ in ($ManualContainer | Sort-Object -Property Name)) {
                                         $PGHASHTABLE = @{}
                                         $PGOBJ.psobject.properties | ForEach-Object { $PGHASHTABLE[$_.Name] = $_.Value }
@@ -197,13 +197,13 @@ function Get-DiagBackupToProtectedGroup {
                                             "<B>Type</B>: $($PGOBJ.Object.Type) <B>Status</B>: $($Status) <B>Schedule</B>: $($PGOBJ.Object.ScheduleOptions.PolicyType)"
                                         )
 
-                                        Convert-TableToHTML -Label $PGOBJ.Name -Name $PGOBJ.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14
+                                        Convert-DiaTableToHTML -Label $PGOBJ.Name -Name $PGOBJ.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14 -IconDebug $IconDebug
 
                                         Edge -From DummyMCContainer -To $PGOBJ.Name @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                                     }
 
                                 } else {
-                                    $Group = Split-Array -inArray ($ManualContainer | Sort-Object -Property Name) -size 2
+                                    $Group = Split-array -inArray ($ManualContainer | Sort-Object -Property Name) -size 2
                                     $Number = 0
                                     while ($Number -ne $Group.Length) {
                                         $Random = Get-Random
@@ -222,7 +222,7 @@ function Get-DiagBackupToProtectedGroup {
                                                     "<B>Type</B>: $($_.Object.Type) <B>Status</B>: $($Status) <B>Schedule</B>: $($_.Object.ScheduleOptions.PolicyType)"
                                                 )
 
-                                                Convert-TableToHTML -Label $_.Name -Name $_.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14
+                                                Convert-DiaTableToHTML -Label $_.Name -Name $_.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14 -IconDebug $IconDebug
                                             }
                                         }
                                         $Number++
@@ -241,10 +241,10 @@ function Get-DiagBackupToProtectedGroup {
                             Edge -From ProtectedGroup -To DummyMCContainer @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                         }
                         if ($IndividualContainer) {
-                            SubGraph ICContainer -Attributes @{Label = (Get-HTMLLabel -Label 'Individual Computers' -IconType "VBR_AGENT_IC" -SubgraphLabel); fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
+                            SubGraph ICContainer -Attributes @{Label = (Get-DiaHTMLLabel -Label 'Individual Computers' -IconType "VBR_AGENT_IC" -ImagesObj $Images -IconDebug $IconDebug -SubgraphLabel); fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
                                 # Node used for subgraph centering
                                 Node DummyICContainer @{Label = 'DummyIC'; style = $SubGraphDebug.style; color = $SubGraphDebug.color; shape = 'plain' }
-                                if ($IndividualContainer.count -le 2) {
+                                if (($IndividualContainer | Measure-Object).count -le 2) {
                                     foreach ($PGOBJ in ($IndividualContainer | Sort-Object -Property Name)) {
                                         $PGHASHTABLE = @{}
                                         $PGOBJ.psobject.properties | ForEach-Object { $PGHASHTABLE[$_.Name] = $_.Value }
@@ -267,12 +267,12 @@ function Get-DiagBackupToProtectedGroup {
                                             $Entities
                                         )
 
-                                        Convert-TableToHTML -Label $PGOBJ.Name -Name $PGOBJ.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14
+                                        Convert-DiaTableToHTML -Label $PGOBJ.Name -Name $PGOBJ.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14 -IconDebug $IconDebug
 
                                         Edge -From DummyICContainer -To $PGOBJ.Name @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                                     }
                                 } else {
-                                    $Group = Split-Array -inArray ($IndividualContainer | Sort-Object -Property Name) -size 2
+                                    $Group = Split-array -inArray ($IndividualContainer | Sort-Object -Property Name) -size 2
                                     $Number = 0
                                     while ($Number -ne $Group.Length) {
                                         $Random = Get-Random
@@ -299,7 +299,7 @@ function Get-DiagBackupToProtectedGroup {
                                                     <br /> $Entities"
                                                 )
 
-                                                Convert-TableToHTML -Label $_.Name -Name $_.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14
+                                                Convert-DiaTableToHTML -Label $_.Name -Name $_.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14 -IconDebug $IconDebug
                                             }
                                         }
                                         $Number++
@@ -318,10 +318,10 @@ function Get-DiagBackupToProtectedGroup {
                             Edge -From ProtectedGroup -To DummyICContainer @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                         }
                         if ($CSVContainer) {
-                            SubGraph CSVContainer -Attributes @{Label = (Get-HTMLLabel -Label 'CSV Computers' -IconType "VBR_AGENT_CSV_Logo" -SubgraphLabel); fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
+                            SubGraph CSVContainer -Attributes @{Label = (Get-DiaHTMLLabel -Label 'CSV Computers' -IconType "VBR_AGENT_CSV_Logo" -ImagesObj $Images -IconDebug $IconDebug -SubgraphLabel); fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
                                 # Node used for subgraph centering
                                 Node DummyCSVContainer @{Label = 'DummyCSVC'; style = $SubGraphDebug.style; color = $SubGraphDebug.color; shape = 'plain' }
-                                if ($CSVContainer.count -le 2) {
+                                if (($CSVContainer | Measure-Object).count -le 2) {
                                     foreach ($PGOBJ in ($CSVContainer | Sort-Object -Property Name)) {
                                         $PGHASHTABLE = @{}
                                         $PGOBJ.psobject.properties | ForEach-Object { $PGHASHTABLE[$_.Name] = $_.Value }
@@ -332,13 +332,13 @@ function Get-DiagBackupToProtectedGroup {
                                             "<B>Credential</B> : $($PGOBJ.Object.Container.MasterCredentials.Name)"
                                         )
 
-                                        Convert-TableToHTML -Label $PGOBJ.Name -Name $PGOBJ.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14
+                                        Convert-DiaTableToHTML -Label $PGOBJ.Name -Name $PGOBJ.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14 -IconDebug $IconDebug
 
                                         Edge -From DummyCSVContainer -To $PGOBJ.Name @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                                     }
 
                                 } else {
-                                    $Group = Split-Array -inArray ($CSVContainer | Sort-Object -Property Name) -size 2
+                                    $Group = Split-array -inArray ($CSVContainer | Sort-Object -Property Name) -size 2
                                     $Number = 0
                                     while ($Number -ne $Group.Length) {
                                         $Random = Get-Random
@@ -353,7 +353,7 @@ function Get-DiagBackupToProtectedGroup {
                                                     "<B>Credential</B> : $($_.Object.Container.MasterCredentials.Name)"
                                                 )
 
-                                                Convert-TableToHTML -Label $_.Name -Name $_.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14
+                                                Convert-DiaTableToHTML -Label $_.Name -Name $_.Name -Row $Rows -HeaderColor "#005f4b" -HeaderFontColor "white" -BorderColor "black" -FontSize 14 -IconDebug $IconDebug
                                             }
                                         }
                                         $Number++