+++ Src/Private/Get-VbrBackupSobrInfo.ps1
@@ -5,7 +5,7 @@ function Get-VbrBackupSobrInfo {
     .DESCRIPTION
         Build a diagram of the configuration of Veeam VBR in PDF/PNG/SVG formats using Psgraph.
     .NOTES
-        Version:        0.5.9
+        Version:        0.6.0
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -63,7 +63,7 @@ function Get-VbrBackupSobrInfo {
                             if (-Not $Sobr.CapacityExtent.Repository.UseGatewayServer) {
                                 Switch ($Sobr.CapacityExtent.Repository.ConnectionType) {
                                     'Gateway' {
-                                        switch ($Sobr.CapacityExtent.Repository.GatewayServer.count) {
+                                        switch (($Sobr.CapacityExtent.Repository.GatewayServer | Measure-Object).count) {
                                             0 { "Disable" }
                                             1 { $Sobr.CapacityExtent.Repository.GatewayServer.Name.Split('.')[0] }
                                             Default { 'Automatic' }
@@ -73,7 +73,7 @@ function Get-VbrBackupSobrInfo {
                                     default { 'Unknown' }
                                 }
                             } else {
-                                switch ($Sobr.CapacityExtent.Repository.GatewayServer.count) {
+                                switch (($Sobr.CapacityExtent.Repository.GatewayServer | Measure-Object).count) {
                                     0 { "Disable" }
                                     1 { $Sobr.CapacityExtent.Repository.GatewayServer.Name.Split('.')[0] }
                                     Default { 'Automatic' }
@@ -88,7 +88,7 @@ function Get-VbrBackupSobrInfo {
                             if (-Not $Sobr.ArchiveExtent.Repository.UseGatewayServer) {
                                 Switch ($Sobr.ArchiveExtent.Repository.GatewayMode) {
                                     'Gateway' {
-                                        switch ($Sobr.ArchiveExtent.Repository.GatewayServer.count) {
+                                        switch (($Sobr.ArchiveExtent.Repository.GatewayServer | Measure-Object).count) {
                                             0 { "Disable" }
                                             1 { $Sobr.ArchiveExtent.Repository.GatewayServer.Name.Split('.')[0] }
                                             Default { 'Automatic' }
@@ -98,7 +98,7 @@ function Get-VbrBackupSobrInfo {
                                     default { 'Unknown' }
                                 }
                             } else {
-                                switch ($Sobr.ArchiveExtent.Repository.GatewayServer.count) {
+                                switch (($Sobr.ArchiveExtent.Repository.GatewayServer | Measure-Object).count) {
                                     0 { "Disable" }
                                     1 { $Sobr.ArchiveExtent.Repository.GatewayServer.Name.Split('.')[0] }
                                     Default { 'Automatic' }
@@ -114,7 +114,7 @@ function Get-VbrBackupSobrInfo {
 
                     $TempSobrInfo = [PSCustomObject]@{
                         Name = "$($Sobr.Name.toUpper())"
-                        Label = Get-NodeIcon -Name "$($Sobr.Name)" -IconType "VBR_SOBR" -Align "Center" -Rows $SobrRows
+                        Label = Get-DiaNodeIcon -Name "$($Sobr.Name)" -IconType "VBR_SOBR" -Align "Center" -Rows $SobrRows -ImagesObj $Images -IconDebug $IconDebug
 
                         Capacity = $Sobr.CapacityExtent.Repository | Select-Object -Property @{Name = 'Name'; Expression = { Remove-SpecialChar -String $_.Name -SpecialChars '\' } }, @{Name = 'Rows'; Expression = { $CapacityRows } }, @{Name = 'Icon'; Expression = { Get-IconType -String $_.Type } }