+++ Src/Private/Get-VbrBackupTapeDrivesInfo.ps1
@@ -5,7 +5,7 @@ function Get-VbrBackupTapeDrivesInfo {
     .DESCRIPTION
         Build a diagram of the configuration of Veeam VBR in PDF/PNG/SVG formats using Psgraph.
     .NOTES
-        Version:        0.5.9
+        Version:        0.6.0
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -41,7 +41,7 @@ function Get-VbrBackupTapeDrivesInfo {
 
                     $TempBackupTapeDriveInfo = [PSCustomObject]@{
                         Name = "$((Remove-SpecialChar -String ($TapeDrive.Name) -SpecialChars '\').toUpper())_$(Get-Random)"
-                        Label = Get-NodeIcon -Name "$((Remove-SpecialChar -String ("Drive $($TapeDrive.Address + 1)").split(".")[0] -SpecialChars '\').toUpper())" -IconType 'VBR_Tape_Drive' -Align "Center" -Rows $Rows
+                        Label = Get-DiaNodeIcon -Name "$((Remove-SpecialChar -String ("Drive $($TapeDrive.Address + 1)").split(".")[0] -SpecialChars '\').toUpper())" -IconType 'VBR_Tape_Drive' -Align "Center" -Rows $Rows -ImagesObj $Images -IconDebug $IconDebug
                         LibraryId = $TapeDrive.LibraryId
                         Id = $TapeDrive.Id
                     }